Andrasove zmeny

+++ src/main/ctollvm/IntegerConstantEvalResult.java
@@ -1,14 +1,14 @@
 package ctollvm;
 public class IntegerConstantEvalResult extends EvalResult {
-  int value = 0;
+  long value = 0;
 
-  public IntegerConstantEvalResult(int value, Type type) {
+  public IntegerConstantEvalResult(long value, Type type) {
     this.value = value;
     this.type = type;
   }
 
   public String getRepresentation() {
-    return Integer.toString(value);
+    return Long.toString(value);
   }
 
   public boolean isLVal() {
